src/ai-mcp-platform/pyproject.toml
[project]
name = "ai-mcp-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncio-mqtt>=0.11.0",
    "fastapi>=0.104.0",
    "fastmcp>=0.2.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "uvicorn>=0.24.0",
]
